Stock Price Predictor.
The Stock Price Predictor Website is a single landing page made with HTML and CSS with an appealing and easy-to-navigate interface that contains links to download a machine learning model to predict the prices of future stocks in the New York Stock Exchange and a group of datasets that show the related information of the data used.

The inspiration for the code for this model was taken from: Author: Aman Kharwal. Title: Stock Price Prediction using Machine Learning. URL: https://thecleverprogrammer.com/2020/11/14/stock-price-prediction-using-machine-learning/

The datasets used in this model were taken from: Author: TS. Title:US historical stock prices with earnings data. URL: https://www.kaggle.com/datasets/tsaustin/us-historical-stock-prices-with-earnings-data